gcp/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "msba405-sample-pipeline"
version = "0.1.0"
description = "NYC Taxi Spark + Snowflake pipeline"
requires-python = ">=3.10"

dependencies = [
    "snowflake-connector-python>=3.0"
]

[tool.setuptools.packages.find]
where = ["."]
